+++ src/index.js
@@ -8,6 +8,7 @@ import isDevMode from 'electron-is-dev';
 import fs from 'fs-extra';
 import path from 'path';
 import windowStateKeeper from 'electron-window-state';
+import { enforceMacOSAppLocation } from 'electron-util';
 
 // Set app directory before loading user modules
 if (process.env.FRANZ_APPDATA_DIR != null) {
@@ -22,7 +23,6 @@ if (isDevMode) {
   app.setPath('userData', path.join(app.getPath('appData'), 'FranzDev'));
 }
 
-
 /* eslint-disable import/first */
 import {
   isMac,
@@ -35,6 +35,7 @@ import Tray from './lib/Tray';
 import Settings from './electron/Settings';
 import handleDeepLink from './electron/deepLinking';
 import { isPositionValid } from './electron/windowUtils';
+import askFormacOSPermissions from './electron/macOSPermissions';
 import { appId } from './package.json'; // eslint-disable-line import/no-unresolved
 import './electron/exception';
 
@@ -277,6 +278,10 @@ const createWindow = () => {
     }
   });
 
+  if (isMac) {
+    askFormacOSPermissions();
+  }
+
   mainWindow.on('show', () => {
     debug('Skip taskbar: false');
     mainWindow.setSkipTaskbar(false);
@@ -313,6 +318,9 @@ if (argv['auth-negotiate-delegate-whitelist']) {
 // initialization and is ready to create browser windows.
 // Some APIs can only be used after this event occurs.
 app.on('ready', () => {
+  // force app to live in /Applications
+  enforceMacOSAppLocation();
+
   // Register App URL
   app.setAsDefaultProtocolClient('franz');