+++ src/stores/ServicesStore.js
@@ -16,6 +16,8 @@ import { gaEvent, statsEvent } from '../lib/analytics';
 import { workspaceStore } from '../features/workspaces';
 import { serviceLimitStore } from '../features/serviceLimit';
 import { RESTRICTION_TYPES } from '../models/Service';
+import { TODOS_RECIPE_ID } from '../features/todos';
+import { SPELLCHECKER_LOCALES } from '../i18n/languages';
 
 const debug = require('debug')('Franz:ServiceStore');
 
@@ -74,6 +76,7 @@ export default class ServicesStore extends Store {
     this.actions.service.openDevToolsForActiveService.listen(this._openDevToolsForActiveService.bind(this));
     this.actions.service.hibernate.listen(this._hibernate.bind(this));
     this.actions.service.awake.listen(this._awake.bind(this));
+    this.actions.service.resetLastPollTimer.listen(this._resetLastPollTimer.bind(this));
 
     this.registerReactions([
       this._focusServiceReaction.bind(this),
@@ -83,6 +86,7 @@ export default class ServicesStore extends Store {
       this._logoutReaction.bind(this),
       this._handleMuteSettings.bind(this),
       this._restrictServiceAccess.bind(this),
+      this._checkForActiveService.bind(this),
     ]);
 
     // Just bind this
@@ -136,20 +140,23 @@ export default class ServicesStore extends Store {
         this._hibernate({ serviceId: service.id });
       }
 
-      if (service.lastPoll && (service.lastPoll) - service.lastPollAnswer > ms('30s')) {
-        // If service did not reply for more than 30s try to reload.
+      if (service.lastPoll && (service.lastPoll - service.lastPollAnswer > ms('1m'))) {
+        // If service did not reply for more than 1m try to reload.
         if (!service.isActive) {
-          if (service.lostRecipeReloadAttempt < 3) {
-            service.webview.reload();
+          if (this.stores.app.isOnline && service.lostRecipeReloadAttempt < 3) {
+            debug(`Reloading service: ${service.name} (${service.id}). Attempt: ${service.lostRecipeReloadAttempt}`);
+            // service.webview.reload();
             service.lostRecipeReloadAttempt += 1;
 
             service.lostRecipeConnection = false;
           }
         } else {
+          debug(`Service lost connection: ${service.name} (${service.id}).`);
           service.lostRecipeConnection = true;
         }
       } else {
         service.lostRecipeConnection = false;
+        service.lostRecipeReloadAttempt = 0;
       }
     });
   }
@@ -208,6 +215,14 @@ export default class ServicesStore extends Store {
     return null;
   }
 
+  @computed get isTodosServiceAdded() {
+    return this.allDisplayed.find(service => service.recipe.id === TODOS_RECIPE_ID && service.isEnabled) || null;
+  }
+
+  @computed get isTodosServiceActive() {
+    return this.active && this.active.recipe.id === TODOS_RECIPE_ID;
+  }
+
   one(id) {
     return this.all.find(service => service.id === id);
   }
@@ -217,10 +232,34 @@ export default class ServicesStore extends Store {
   }
 
   // Actions
-  @action async _createService({ recipeId, serviceData, redirect = true }) {
+  async _createService({
+    recipeId, serviceData, redirect = true, skipCleanup = false,
+  }) {
     if (serviceLimitStore.userHasReachedServiceLimit) return;
 
-    const data = this._cleanUpTeamIdAndCustomUrl(recipeId, serviceData);
+    if (!this.stores.recipes.isInstalled(recipeId)) {
+      debug(`Recipe "${recipeId}" is not installed, installing recipe`);
+      await this.stores.recipes._install({ recipeId });
+      debug(`Recipe "${recipeId}" installed`);
+    }
+
+    // set default values for serviceData
+    Object.assign({
+      isEnabled: true,
+      isHibernationEnabled: false,
+      isNotificationEnabled: true,
+      isBadgeEnabled: true,
+      isMuted: false,
+      customIcon: false,
+      isDarkModeEnabled: false,
+      spellcheckerLanguage: SPELLCHECKER_LOCALES[this.stores.settings.app.spellcheckerLanguage],
+    }, serviceData);
+
+    let data = serviceData;
+
+    if (!skipCleanup) {
+      data = this._cleanUpTeamIdAndCustomUrl(recipeId, serviceData);
+    }
 
     const response = await this.createServiceRequest.execute(recipeId, data)._promise;
 
@@ -357,6 +396,10 @@ export default class ServicesStore extends Store {
     service.isActive = true;
     this._awake({ serviceId: service.id });
     service.lastUsed = Date.now();
+
+    if (this.active.recipe.id === TODOS_RECIPE_ID && !this.stores.todos.settings.isFeatureEnabledByUser) {
+      this.actions.todos.toggleTodosFeatureVisibility();
+    }
 
     statsEvent('activate-service', service.recipe.id);
 
@@ -563,7 +606,11 @@ export default class ServicesStore extends Store {
     service.resetMessageCount();
     service.lostRecipeConnection = false;
 
-    service.webview.loadURL(service.url);
+    if (service.recipe.id === TODOS_RECIPE_ID) {
+      return this.actions.todos.reload();
+    }
+
+    return service.webview.loadURL(service.url);
   }
 
   @action _reloadActive() {
@@ -648,15 +695,18 @@ export default class ServicesStore extends Store {
 
   @action _openDevTools({ serviceId }) {
     const service = this.one(serviceId);
-
-    service.webview.openDevTools();
+    if (service.recipe.id === TODOS_RECIPE_ID) {
+      this.actions.todos.openDevTools();
+    } else {
+      service.webview.openDevTools();
+    }
   }
 
   @action _openDevToolsForActiveService() {
     const service = this.active;
 
     if (service) {
-      service.webview.openDevTools();
+      this._openDevTools({ serviceId: service.id });
     } else {
       debug('No service is active');
     }
@@ -678,6 +728,24 @@ export default class ServicesStore extends Store {
     const service = this.one(serviceId);
     service.isHibernating = false;
     service.liveFrom = Date.now();
+  }
+
+  @action _resetLastPollTimer({ serviceId = null }) {
+    debug(`Reset last poll timer for ${serviceId ? `service: "${serviceId}"` : 'all services'}`);
+
+    const resetTimer = (service) => {
+      service.lastPollAnswer = Date.now();
+      service.lastPoll = Date.now();
+    };
+
+    if (!serviceId) {
+      this.allDisplayed.forEach(service => resetTimer(service));
+    } else {
+      const service = this.one(serviceId);
+      if (service) {
+        resetTimer(service);
+      }
+    }
   }
 
   // Reactions
@@ -805,6 +873,18 @@ export default class ServicesStore extends Store {
 
       return service;
     });
+  }
+
+  _checkForActiveService() {
+    if (this.stores.router.location.pathname.includes('auth/signup')) {
+      return;
+    }
+
+    if (this.allDisplayed.findIndex(service => service.isActive) === -1 && this.allDisplayed.length !== 0) {
+      debug('No active service found, setting active service to index 0');
+
+      this._setActive({ serviceId: this.allDisplayed[0].id });
+    }
   }
 
   // Helper